Rename storybook internal components directory
