+++ react/stories/__components__/ColorSwatch/index.js
