+++ react/stories/__components__/TypeSpecimen/index.js
